Simplify panel texture attributes, add some helpful comments

+++ Orbitersdk/samples/ProjectApollo/src_lm_new/LM.h
@@ -87,8 +87,11 @@ private:
 
 	// Handle for new 2D Panel mesh
 	MESHHANDLE hPanelMesh;
-	static std::array<SURFHANDLE, (int)PANEL_ID::panelCount> panelTextures;	// handles for NEW 2D Panel textures
-	static std::array<RECT, (int)PANEL_ID::panelCount> panelDimensions;		// rectangles for NEW 2D Panel dimensions, since textures may be larger
-	static std::array<SURFHANDLE, (int)PANELOBJECT_ID::objectCount> panelObjectTextures;	// handles for NEW 2D Panel object textures
-	static std::array<RECT, (int)PANELOBJECT_ID::objectCount> panelObjectDimensions;		// rectangles for NEW 2D Panel object dimensions, since textures may be larger
+	// For clarification, the "dimensions" are the locations on the panel where a given texture/surface will be placed.
+	// To obtain the actual surface dimensions (which may include blank space due to DDS textures needing their resolution
+	// to be a power of two, use the oapiGetSurfaceSize() function.
+	static std::array<SURFHANDLE, PANEL_ID::panelCount> panelTextures;	// handles for NEW 2D Panel textures
+	static std::array<RECT, PANEL_ID::panelCount> panelDimensions;		// rectangles for NEW 2D Panel dimensions, since textures may be larger
+	static std::array<SURFHANDLE, PANELOBJECT_ID::objectCount> panelObjectTextures;	// handles for NEW 2D Panel object textures
+	static std::array<RECT, PANELOBJECT_ID::objectCount> panelObjectDimensions;		// rectangles for NEW 2D Panel object dimensions, since textures may be larger
 };
